+++ .storybook/main.js
@@ -1,11 +1,19 @@
 const path = require('path');
 
 module.exports = {
-  addons: ['@storybook/addon-docs', '@storybook/addon-knobs/register', '@storybook/addon-actions/register'],
   stories: ['../src/stories/*.stories.ts'],
+  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
+  core: {
+    builder: 'webpack5'
+  },
   webpackFinal: async (config, { configType }) => {
     config.module.rules.push({
-      test: /\.(ts|tsx)$/,
+      test: /\.glsl$/,
+      use: ['raw-loader']
+    });
+
+    config.module.rules.push({
+      test: /\.(tsx?)$/,
       use: [
         {
           loader: require.resolve('ts-loader'),
@@ -15,7 +23,13 @@ module.exports = {
         }
       ]
     });
-    config.resolve.extensions.push('.ts', '.tsx');
+
+    const sourceLoader = config.module.rules.findIndex((r) => r.loader && r.loader.includes('source-loader'));
+
+    if (sourceLoader > -1) {
+      // TODO: Investigate why source-loader is messing with graphics.add(string, object); expressions
+      config.module.rules.splice(sourceLoader, 1);
+    }
 
     if (configType === 'PRODUCTION') {
       config.mode = 'development';