+++ src/engine/Actions/Action/Delay.ts
@@ -1,16 +1,11 @@
-import { Actor } from '../../Actor';
 import { Action } from '../Action';
 
 export class Delay implements Action {
-  public x: number;
-  public y: number;
-  private _actor: Actor;
   private _elapsedTime: number = 0;
   private _delay: number;
   private _started: boolean = false;
   private _stopped = false;
-  constructor(actor: Actor, delay: number) {
-    this._actor = actor;
+  constructor(delay: number) {
     this._delay = delay;
   }
 
@@ -18,9 +13,6 @@ export class Delay implements Action {
     if (!this._started) {
       this._started = true;
     }
-
-    this.x = this._actor.pos.x;
-    this.y = this._actor.pos.y;
 
     this._elapsedTime += delta;
   }