+++ src/engine/Actions/Action/Meet.ts
@@ -1,10 +1,14 @@
-import { Actor } from '../../Actor';
+import { MotionComponent } from '../../EntityComponentSystem/Components/MotionComponent';
+import { TransformComponent } from '../../EntityComponentSystem/Components/TransformComponent';
+import { Entity } from '../../EntityComponentSystem/Entity';
 import { Vector, vec } from '../../Math/vector';
 import { Action } from '../Action';
 
 export class Meet implements Action {
-  private _actor: Actor;
-  private _actorToMeet: Actor;
+  private _tx: TransformComponent;
+  private _motion: MotionComponent;
+  private _meetTx: TransformComponent;
+  private _meetMotion: MotionComponent;
   public x: number;
   public y: number;
   private _current: Vector;
@@ -16,11 +20,13 @@ export class Meet implements Action {
   private _stopped = false;
   private _speedWasSpecified = false;
 
-  constructor(actor: Actor, actorToMeet: Actor, speed?: number) {
-    this._actor = actor;
-    this._actorToMeet = actorToMeet;
-    this._current = new Vector(this._actor.pos.x, this._actor.pos.y);
-    this._end = new Vector(actorToMeet.pos.x, actorToMeet.pos.y);
+  constructor(actor: Entity, actorToMeet: Entity, speed?: number) {
+    this._tx = actor.get(TransformComponent);
+    this._motion = actor.get(MotionComponent);
+    this._meetTx = actorToMeet.get(TransformComponent);
+    this._meetMotion = actorToMeet.get(MotionComponent);
+    this._current = new Vector(this._tx.pos.x, this._tx.pos.y);
+    this._end = new Vector(this._meetTx.pos.x, this._meetTx.pos.y);
     this._speed = speed || 0;
 
     if (speed !== undefined) {
@@ -35,22 +41,22 @@ export class Meet implements Action {
       this._dir = this._end.sub(this._current).normalize();
     }
 
-    const actorToMeetSpeed = Math.sqrt(Math.pow(this._actorToMeet.vel.x, 2) + Math.pow(this._actorToMeet.vel.y, 2));
+    const actorToMeetSpeed = Math.sqrt(Math.pow(this._meetMotion.vel.x, 2) + Math.pow(this._meetMotion.vel.y, 2));
     if (actorToMeetSpeed !== 0 && !this._speedWasSpecified) {
       this._speed = actorToMeetSpeed;
     }
-    this._current = vec(this._actor.pos.x, this._actor.pos.y);
+    this._current = vec(this._tx.pos.x, this._tx.pos.y);
 
-    this._end = vec(this._actorToMeet.pos.x, this._actorToMeet.pos.y);
+    this._end = vec(this._meetTx.pos.x, this._meetTx.pos.y);
     this._distanceBetween = this._current.distance(this._end);
     this._dir = this._end.sub(this._current).normalize();
 
     const m = this._dir.scale(this._speed);
-    this._actor.vel = vec(m.x, m.y);
+    this._motion.vel = vec(m.x, m.y);
 
     if (this.isComplete()) {
-      this._actor.pos = vec(this._end.x, this._end.y);
-      this._actor.vel = vec(0, 0);
+      this._tx.pos = vec(this._end.x, this._end.y);
+      this._motion.vel = vec(0, 0);
     }
   }
 
@@ -59,7 +65,7 @@ export class Meet implements Action {
   }
 
   public stop(): void {
-    this._actor.vel = vec(0, 0);
+    this._motion.vel = vec(0, 0);
     this._stopped = true;
   }