+++ src/engine/Actions/Action/RotateTo.ts
@@ -1,10 +1,13 @@
-import { Actor } from '../../Actor';
 import { Action } from '../Action';
 import { RotationType } from '../RotationType';
 import * as Util from '../../Util/Util';
+import { TransformComponent } from '../../EntityComponentSystem/Components/TransformComponent';
+import { MotionComponent } from '../../EntityComponentSystem/Components/MotionComponent';
+import { Entity } from '../../EntityComponentSystem/Entity';
 
 export class RotateTo implements Action {
-  private _actor: Actor;
+  private _tx: TransformComponent;
+  private _motion: MotionComponent;
   public x: number;
   public y: number;
   private _start: number;
@@ -19,8 +22,9 @@ export class RotateTo implements Action {
   private _currentNonCannonAngle: number;
   private _started = false;
   private _stopped = false;
-  constructor(actor: Actor, angleRadians: number, speed: number, rotationType?: RotationType) {
-    this._actor = actor;
+  constructor(entity: Entity, angleRadians: number, speed: number, rotationType?: RotationType) {
+    this._tx = entity.get(TransformComponent);
+    this._motion = entity.get(MotionComponent);
     this._end = angleRadians;
     this._speed = speed;
     this._rotationType = rotationType || RotationType.ShortestPath;
@@ -29,8 +33,8 @@ export class RotateTo implements Action {
   public update(_delta: number): void {
     if (!this._started) {
       this._started = true;
-      this._start = this._actor.rotation;
-      this._currentNonCannonAngle = this._actor.rotation;
+      this._start = this._tx.rotation;
+      this._currentNonCannonAngle = this._tx.rotation;
       const distance1 = Math.abs(this._end - this._start);
       const distance2 = Util.TwoPI - distance1;
       if (distance1 > distance2) {
@@ -79,12 +83,12 @@ export class RotateTo implements Action {
       }
     }
 
-    this._actor.angularVelocity = this._direction * this._speed;
+    this._motion.angularVelocity = this._direction * this._speed;
     this._currentNonCannonAngle += this._direction * this._speed * (_delta / 1000);
 
     if (this.isComplete()) {
-      this._actor.rotation = this._end;
-      this._actor.angularVelocity = 0;
+      this._tx.rotation = this._end;
+      this._motion.angularVelocity = 0;
       this._stopped = true;
     }
   }
@@ -95,7 +99,7 @@ export class RotateTo implements Action {
   }
 
   public stop(): void {
-    this._actor.angularVelocity = 0;
+    this._motion.angularVelocity = 0;
     this._stopped = true;
   }